fix include conflict & ccl build rpath error
